+++ backend/cwa_lib/pydantic_schemas/su_manage_vdbs.py
@@ -1,5 +1,6 @@
-from pydantic import BaseModel
+from pydantic import BaseModel, field_validator
 from .generic_table import TableQueryResult
+from common.features.gvdbs_retr_params import GVDBsDefRetrParams
 
 
 class SuManageVDBsRead(BaseModel):
@@ -11,6 +12,7 @@ class SuManageVDBsRead(BaseModel):
     gvdbs_name: str
     gvdbs_url: str
     gvdbs_collection: str
+    gvdbs_retr_params: str
     gvdbs_status: str
     gvdbs_status_text: str
     class Config:
@@ -27,7 +29,7 @@ class SuManageVDBsCreate(BaseModel):
     gvdbs_type: str
     gvdbs_name: str
     gvdbs_url: str
-    gvdbs_collection: str
+    gvdbs_collection: str    
 
 
 class SuManageVDBsUpdate(BaseModel):
@@ -39,3 +41,10 @@ class SuManageVDBsUpdate(BaseModel):
     gvdbs_name: str | None = None
     gvdbs_url: str | None = None
     gvdbs_collection: str | None = None
+    gvdbs_retr_params: str | None = None
+    @field_validator('gvdbs_retr_params')
+    @staticmethod
+    def validate__gvdbs_retr_params(v: str | None) -> str | None:
+        if v is not None:
+            GVDBsDefRetrParams.from_dict(v)
+        return v